+++ source/modules/gui/widgets/label.cpp
@@ -16,7 +16,7 @@ namespace gui
 		this->text = font::Text(master.glContext, labelTemplate.text.text, mFont, textPosition, labelTemplate.text.color, labelTemplate.text.size);
 
 		if (size == gml::Vec2f(0.0f)) {
-			prefSize = this->text.getSize() + gml::Vec2f(this->borderSize) * 2 + padding * 2;
+			prefSize = this->text.getSize() + gml::Vec2f((float)this->borderSize) * 2 + padding * 2;
 		}
 		else {
 			prefSize = size;
@@ -28,7 +28,7 @@ namespace gui
 	void Label::setText(TextTemplate text)
 	{
 		font::Font & mFont = master.fontManager->getFont(text.font);
-		gml::Vec2f textPosition = this->position + gml::Vec2f(borderSize) + padding;
+		gml::Vec2f textPosition = this->position + gml::Vec2f((float)borderSize) + padding;
 
 		this->text = font::Text(master.glContext, text.text, mFont, textPosition, text.color, text.size);
 
@@ -59,7 +59,7 @@ namespace gui
 	void Label::updateGeometry()
 	{
 		Widget::updateGeometry();
-		gml::Vec2f textPosition = position + gml::Vec2f(borderSize) + padding;
+		gml::Vec2f textPosition = position + gml::Vec2f((float)borderSize) + padding;
 		text.setPosition(textPosition);
 	}
 }
